requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
openai==1.10.0
aiosqlite==0.19.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
